requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0
loguru==0.7.2
email-validator==2.1.1
aiosqlite==0.19.0
websockets==12.0
requests==2.31.0
